+++ projects/packages/sync/src/class-sender.php
@@ -292,6 +292,7 @@ class Sender {
 	 */
 	public function do_full_sync() {
 		$sync_module = Modules::get_module( 'full-sync' );
+		'@phan-var Modules\Full_Sync_Immediately|Modules\Full_Sync $sync_module';
 		if ( ! $sync_module ) {
 			return;
 		}
@@ -313,7 +314,7 @@ class Sender {
 
 		$this->continue_full_sync_enqueue();
 		// immediate full sync sends data in continue_full_sync_enqueue.
-		if ( ! str_contains( get_class( $sync_module ), 'Full_Sync_Immediately' ) ) {
+		if ( ! $sync_module instanceof Modules\Full_Sync_Immediately ) {
 			return $this->do_sync_and_set_delays( $this->full_sync_queue );
 		} else {
 			$status = $sync_module->get_status();
@@ -342,7 +343,9 @@ class Sender {
 			return false;
 		}
 
-		Modules::get_module( 'full-sync' )->continue_enqueuing();
+		$full_sync_module = Modules::get_module( 'full-sync' );
+		'@phan-var Modules\Full_Sync_Immediately|Modules\Full_Sync $full_sync_module';
+		$full_sync_module->continue_enqueuing();
 
 		$this->set_next_sync_time( time() + $this->get_enqueue_wait_time(), 'full-sync-enqueue' );
 	}