+++ projects/plugins/jetpack/extensions/blocks/ai-assistant/inline-extensions/with-ai-extension.tsx
@@ -8,16 +8,16 @@ import {
 } from '@automattic/jetpack-ai-client';
 import { BlockControls, useBlockProps } from '@wordpress/block-editor';
 import { createHigherOrderComponent } from '@wordpress/compose';
-import { dispatch, select, useDispatch } from '@wordpress/data';
-import { useCallback, useEffect, useState, useRef } from '@wordpress/element';
+import { useDispatch, useSelect } from '@wordpress/data';
+import { useCallback, useEffect, useState, useRef, useMemo } from '@wordpress/element';
 import { addFilter } from '@wordpress/hooks';
 import debugFactory from 'debug';
 import React from 'react';
 /*
  * Internal dependencies
  */
-import { EXTENDED_INLINE_BLOCKS } from '../extensions/ai-assistant';
-import { BuildPromptOptionsProps, buildPromptForExtensions } from '../lib/prompt';
+import useAiFeature from '../hooks/use-ai-feature';
+import { mapInternalPromptTypeToBackendPromptType } from '../lib/prompt/backend-prompt';
 import { blockHandler } from './block-handler';
 import AiAssistantInput from './components/ai-assistant-input';
 import AiAssistantExtensionToolbarDropdown from './components/ai-assistant-toolbar-dropdown';
@@ -34,60 +34,106 @@ import type { PromptTypeProp } from '../lib/prompt';
 
 const debug = debugFactory( 'jetpack-ai-assistant:extensions:with-ai-extension' );
 
-// HOC to populate the block's edit component with the AI Assistant bar and button.
+const blockExtensionMapper = {
+	'core/heading': 'heading',
+};
+
+// Defines where the block controls should be placed in the toolbar
+const blockControlsProps = {
+	group: 'block' as const,
+};
+
+type RequestOptions = {
+	promptType: PromptTypeProp;
+	options?: AiAssistantDropdownOnChangeOptionsArgProps;
+	humanText?: string;
+};
+
+type CoreEditorDispatch = { undo: () => Promise< void > };
+type CoreEditorSelect = { getCurrentPostId: () => number };
+
+// HOC to populate the block's edit component with the AI Assistant control inpuit and toolbar button.
 const blockEditWithAiComponents = createHigherOrderComponent( BlockEdit => {
 	return props => {
 		const { clientId, isSelected, name: blockName } = props;
+		// Ref to the control wrapper, its height and its ResizeObserver, for positioning adjustments.
 		const controlRef: React.MutableRefObject< HTMLDivElement | null > = useRef( null );
 		const controlHeight = useRef< number >( 0 );
-		const inputRef: React.MutableRefObject< HTMLInputElement | null > = useRef( null );
 		const controlObserver = useRef< ResizeObserver | null >( null );
+		// Ref to the original block style to reset it when the AI Control is closed.
 		const blockStyle = useRef< string >( '' );
+		// Ref to the input element to focus on it when the AI Control is displayed or when a request is done.
+		// Also used to determine the ownerDocument, as the editor can be in an iframe.
+		const inputRef: React.MutableRefObject< HTMLInputElement | null > = useRef( null );
 		const ownerDocument = useRef< Document >( document );
+		// A human-readable action to be displayed in the input when a toolbar suggestion is requested, like "Translate: Japanese".
 		const [ action, setAction ] = useState< string >( '' );
-		const [ requestCount, setRequestCount ] = useState( 0 );
-
-		// Only extend the allowed block types.
-		const possibleToExtendBlock = isPossibleToExtendBlock( {
-			blockName,
-			clientId,
-		} );
-
+		// Count of consecutive successful AI Assistant requests for the Undo button.
+		const [ consecutiveRequestCount, setConsecutiveRequestCount ] = useState( 0 );
+		// The last request made by the user, to be used when the user clicks the "Try Again" button.
+		const [ lastRequest, setLastRequest ] = useState< RequestOptions | null >( null );
+		// State to display the AI Control or not.
 		const [ showAiControl, setShowAiControl ] = useState( false );
+		// Data and functions from the editor.
+		const { undo } = useDispatch( 'core/editor' ) as CoreEditorDispatch;
+		const { postId } = useSelect( select => {
+			const { getCurrentPostId } = select( 'core/editor' ) as CoreEditorSelect;
 
-		const { getCurrentPostId } = select( 'core/editor' );
-		const postId = getCurrentPostId();
-
-		const { increaseAiAssistantRequestsCount } = useDispatch( 'wordpress-com/plans' );
-
-		const onDone = useCallback( () => {
-			increaseAiAssistantRequestsCount();
-			setRequestCount( count => count + 1 );
-		}, [ increaseAiAssistantRequestsCount ] );
-
-		const onError = useCallback(
-			error => {
-				// Increase the AI Suggestion counter only for valid errors.
-				if ( error.code === ERROR_NETWORK || error.code === ERROR_QUOTA_EXCEEDED ) {
-					return;
-				}
-
-				increaseAiAssistantRequestsCount();
-			},
-			[ increaseAiAssistantRequestsCount ]
-		);
-
+			return { postId: getCurrentPostId() };
+		}, [] );
+		// The block's id to find it in the DOM for the positioning adjustments.
 		const { id } = useBlockProps();
+		// Jetpack AI Assistant feature functions.
+		const { increaseRequestsCount, dequeueAsyncRequest } = useAiFeature();
 
 		// Data and functions with block-specific implementations.
-		const { onSuggestion: onBlockSuggestion, getContent } = blockHandler( blockName, clientId );
+		const { onSuggestion: onBlockSuggestion, getContent } = useMemo(
+			() => blockHandler( blockName, clientId ),
+			[ blockName, clientId ]
+		);
 
+		// Called when the user clicks the "Ask AI Assistant" button.
+		const handleAskAiAssistant = useCallback( () => {
+			setShowAiControl( true );
+		}, [] );
+
+		// Function to get the messages array for the request.
+		const getRequestMessages = useCallback(
+			( {
+				promptType,
+				options,
+			}: {
+				promptType: PromptTypeProp;
+				options?: AiAssistantDropdownOnChangeOptionsArgProps;
+			} ) => {
+				const blockContent = getContent();
+
+				const extension = blockExtensionMapper[ blockName ];
+
+				return [
+					{
+						role: 'jetpack-ai' as const,
+						context: {
+							type: mapInternalPromptTypeToBackendPromptType( promptType, extension ),
+							content: blockContent,
+							request: options?.userPrompt,
+							tone: options?.tone,
+							language: options?.language,
+						},
+					},
+				];
+			},
+			[ blockName, getContent ]
+		);
+
+		// Called when a suggestion chunk is received.
 		const onSuggestion = useCallback(
 			( suggestion: string ) => {
 				onBlockSuggestion( suggestion );
 
 				// Make sure the block element has the necessary bottom padding, as it can be replaced or changed
 				const block = ownerDocument.current.getElementById( id );
+
 				if ( block && controlRef.current ) {
 					block.style.paddingBottom = `${ controlHeight.current + 16 }px`;
 				}
@@ -95,12 +141,37 @@ const blockEditWithAiComponents = createHigherOrderComponent( BlockEdit => {
 			[ id, onBlockSuggestion ]
 		);
 
+		// Called after the last suggestion chunk is received.
+		const onDone = useCallback( () => {
+			increaseRequestsCount();
+			setConsecutiveRequestCount( count => count + 1 );
+			inputRef.current?.focus();
+			setAction( '' );
+			setLastRequest( null );
+		}, [ increaseRequestsCount ] );
+
+		// Called when an error is received.
+		const onError = useCallback(
+			error => {
+				setAction( '' );
+
+				debug( 'Request error', error );
+
+				// Increase the AI Suggestion counter only for valid errors.
+				if ( error.code === ERROR_NETWORK || error.code === ERROR_QUOTA_EXCEEDED ) {
+					return;
+				}
+
+				increaseRequestsCount();
+			},
+			[ increaseRequestsCount ]
+		);
+
 		const {
 			request,
 			stopSuggestion,
 			requestingState,
 			error,
-			suggestion,
 			reset: resetSuggestions,
 		} = useAiSuggestions( {
 			onSuggestion,
@@ -112,6 +183,88 @@ const blockEditWithAiComponents = createHigherOrderComponent( BlockEdit => {
 			},
 		} );
 
+		// Called when a suggestion from the toolbar is requested, like "Change tone".
+		const handleRequestSuggestion = useCallback< OnRequestSuggestion >(
+			( promptType, options, humanText ) => {
+				setShowAiControl( true );
+
+				if ( humanText ) {
+					setAction( humanText );
+				}
+
+				const messages = getRequestMessages( { promptType, options } );
+
+				debug( 'Request suggestion', promptType, options );
+
+				setLastRequest( { promptType, options, humanText } );
+
+				/*
+				 * Always dequeue/cancel the AI Assistant feature async request,
+				 * in case there is one pending,
+				 * when performing a new AI suggestion request.
+				 */
+				dequeueAsyncRequest();
+
+				request( messages );
+			},
+			[ dequeueAsyncRequest, getRequestMessages, request ]
+		);
+
+		// Called when the user types a custom prompt.
+		const handleUserRequest = useCallback(
+			( userPrompt: string ) => {
+				const promptType = 'userPrompt';
+				const options = { userPrompt };
+
+				handleRequestSuggestion( promptType, options );
+			},
+			[ handleRequestSuggestion ]
+		);
+
+		// Called when the user clicks the "Stop" button in the input.
+		const handleStopSuggestion = useCallback( () => {
+			stopSuggestion();
+
+			inputRef.current?.focus();
+		}, [ stopSuggestion ] );
+
+		// Called when the user clicks the "Try Again" button in the input error message.
+		const handleTryAgain = useCallback( () => {
+			if ( lastRequest ) {
+				handleRequestSuggestion(
+					lastRequest.promptType,
+					lastRequest.options,
+					lastRequest.humanText
+				);
+			}
+		}, [ lastRequest, handleRequestSuggestion ] );
+
+		// Cleanup function.
+		const handleClose = useCallback( () => {
+			setShowAiControl( false );
+			resetSuggestions();
+			setAction( '' );
+			setConsecutiveRequestCount( 0 );
+			setLastRequest( null );
+		}, [ resetSuggestions ] );
+
+		// Called when the user clicks the "Undo" button after a successful request.
+		const handleUndo = useCallback( async () => {
+			for ( let i = 0; i < consecutiveRequestCount; i++ ) {
+				await undo();
+			}
+
+			handleClose();
+		}, [ consecutiveRequestCount, handleClose, undo ] );
+
+		// Closes the AI Control if the block is deselected.
+		useEffect( () => {
+			if ( ! isSelected ) {
+				handleClose();
+			}
+		}, [ isSelected, handleClose ] );
+
+		// Focus the input when the AI Control is displayed and set the ownerDocument.
 		useEffect( () => {
 			if ( inputRef.current ) {
 				// Save the block's ownerDocument to use it later, as the editor can be in an iframe.
@@ -121,6 +274,8 @@ const blockEditWithAiComponents = createHigherOrderComponent( BlockEdit => {
 			}
 		}, [ showAiControl ] );
 
+		// Adjusts the input position in the editor by increasing the block's bottom-padding
+		// and setting the control's margin-top, "wrapping" the input with the block.
 		useEffect( () => {
 			let block = ownerDocument.current.getElementById( id );
 
@@ -154,90 +309,13 @@ const blockEditWithAiComponents = createHigherOrderComponent( BlockEdit => {
 				controlObserver.current = null;
 				controlHeight.current = 0;
 			}
-		}, [ clientId, controlObserver, id, showAiControl ] );
 
-		// Only extend the target block.
-		if ( ! possibleToExtendBlock ) {
-			return <BlockEdit { ...props } />;
-		}
-
-		// Defines where the block controls should be placed in the toolbar
-		const blockControlsProps = {
-			group: 'block' as const,
-		};
-
-		const onAskAiAssistant = () => {
-			setShowAiControl( true );
-		};
-
-		const getRequestMessages = ( {
-			promptType,
-			options,
-			userPrompt,
-		}: {
-			promptType: PromptTypeProp;
-			options?: AiAssistantDropdownOnChangeOptionsArgProps;
-			userPrompt?: string;
-		} ) => {
-			const blockContent = getContent();
-
-			const promptOptions: BuildPromptOptionsProps = {
-				tone: options?.tone,
-				language: options?.language,
-				fromExtension: true,
+			return () => {
+				if ( controlObserver.current ) {
+					controlObserver.current.disconnect();
+				}
 			};
-
-			return buildPromptForExtensions( {
-				blockContent,
-				options: promptOptions,
-				type: promptType,
-				userPrompt,
-			} );
-		};
-
-		const onRequestSuggestion: OnRequestSuggestion = ( promptType, options, humanText ) => {
-			setShowAiControl( true );
-
-			if ( humanText ) {
-				setAction( humanText );
-			}
-
-			const messages = getRequestMessages( { promptType, options } );
-
-			debug( 'onRequestSuggestion', promptType, options );
-
-			request( messages );
-		};
-
-		const onClose = useCallback( () => {
-			setShowAiControl( false );
-			resetSuggestions();
-			setAction( '' );
-			setRequestCount( 0 );
-		}, [ resetSuggestions ] );
-
-		const onUserRequest = ( userPrompt: string ) => {
-			const promptType = 'userPrompt';
-			const options = {};
-			const messages = getRequestMessages( { promptType, options, userPrompt } );
-
-			request( messages );
-		};
-
-		// Close the AI Control if the block is deselected.
-		useEffect( () => {
-			if ( ! isSelected ) {
-				onClose();
-			}
-		}, [ isSelected, onClose ] );
-
-		const onUndo = async () => {
-			for ( let i = 0; i < requestCount; i++ ) {
-				await dispatch( 'core/editor' ).undo();
-			}
-
-			onClose();
-		};
+		}, [ clientId, controlObserver, id, showAiControl ] );
 
 		return (
 			<>
@@ -245,26 +323,25 @@ const blockEditWithAiComponents = createHigherOrderComponent( BlockEdit => {
 
 				{ showAiControl && (
 					<AiAssistantInput
-						clientId={ clientId }
-						postId={ postId }
 						requestingState={ requestingState }
 						requestingError={ error }
-						suggestion={ suggestion }
 						wrapperRef={ controlRef }
 						inputRef={ inputRef }
 						action={ action }
-						request={ onUserRequest }
-						stopSuggestion={ stopSuggestion }
-						close={ onClose }
-						undo={ onUndo }
+						blockType={ blockName }
+						request={ handleUserRequest }
+						stopSuggestion={ handleStopSuggestion }
+						close={ handleClose }
+						undo={ handleUndo }
+						tryAgain={ handleTryAgain }
 					/>
 				) }
 
 				<BlockControls { ...blockControlsProps }>
 					<AiAssistantExtensionToolbarDropdown
 						blockType={ blockName }
-						onAskAiAssistant={ onAskAiAssistant }
-						onRequestSuggestion={ onRequestSuggestion }
+						onAskAiAssistant={ handleAskAiAssistant }
+						onRequestSuggestion={ handleRequestSuggestion }
 					/>
 				</BlockControls>
 			</>
@@ -280,8 +357,10 @@ const blockEditWithAiComponents = createHigherOrderComponent( BlockEdit => {
  * @returns {object}          The extended block settings.
  */
 function blockWithInlineExtension( settings, name: ExtendedInlineBlockProp ) {
-	// Only extend the allowed block types.
-	if ( ! EXTENDED_INLINE_BLOCKS.includes( name ) ) {
+	// Only extend the allowed block types and when AI is enabled
+	const possibleToExtendBlock = isPossibleToExtendBlock( name );
+
+	if ( ! possibleToExtendBlock ) {
 		return settings;
 	}