+++ projects/plugins/jetpack/modules/sso/class.jetpack-sso-user-admin.php
@@ -103,12 +103,18 @@ if ( ! class_exists( 'Jetpack_SSO_User_Admin' ) ) :
 					$event    = 'sso_user_invite_revoked';
 
 					if ( 200 !== wp_remote_retrieve_response_code( $response ) ) {
+						$body                = json_decode( wp_remote_retrieve_body( $response ) );
+						$tracking_event_data = array(
+							'success'    => 'false',
+							'error_code' => 'invalid-revoke-api-error',
+						);
+
+						if ( ! empty( $body ) && ! empty( $body->message ) ) {
+							$tracking_event_data['error_message'] = $body->message;
+						}
 						self::$tracking->record_user_event(
 							$event,
-							array(
-								'success'       => 'false',
-								'error_message' => 'invalid-revoke-api-error',
-							)
+							$tracking_event_data
 						);
 						return $response;
 					}
@@ -160,6 +166,10 @@ if ( ! class_exists( 'Jetpack_SSO_User_Admin' ) ) :
 
 			if ( $_GET['jetpack-sso-invite-user'] === 'successful-revoke' ) {
 				return wp_admin_notice( __( 'User invite revoked successfully.', 'jetpack' ), array( 'type' => 'success' ) );
+			}
+
+			if ( $_GET['jetpack-sso-invite-user'] === 'failed' && isset( $_GET['jetpack-sso-api-error-message'] ) ) {
+				return wp_admin_notice( wp_kses( wp_unslash( $_GET['jetpack-sso-api-error-message'] ), array() ), array( 'type' => 'error' ) );
 			}
 
 			if ( $_GET['jetpack-sso-invite-user'] === 'failed' && isset( $_GET['jetpack-sso-invite-error'] ) ) {
@@ -252,19 +262,27 @@ if ( ! class_exists( 'Jetpack_SSO_User_Admin' ) ) :
 				);
 
 				if ( 200 !== wp_remote_retrieve_response_code( $response ) ) {
-					$error        = 'invalid-invite-api-error';
+					$error_code   = 'invalid-invite-api-error';
 					$query_params = array(
 						'jetpack-sso-invite-user'  => 'failed',
-						'jetpack-sso-invite-error' => $error,
+						'jetpack-sso-invite-error' => $error_code,
 						'_wpnonce'                 => $nonce,
 					);
 
+					$tracking_event_data = array(
+						'success'    => 'false',
+						'error_code' => $error_code,
+					);
+
+					$body = json_decode( wp_remote_retrieve_body( $response ) );
+					if ( ! empty( $body ) && ! empty( $body->message ) ) {
+						$query_params['jetpack-sso-api-error-message'] = $body->message;
+						$tracking_event_data['error_message']          = $body->message;
+					}
+
 					self::$tracking->record_user_event(
 						$event,
-						array(
-							'success'       => 'false',
-							'error_message' => $error,
-						)
+						$tracking_event_data
 					);
 					return self::create_error_notice_and_redirect( $query_params );
 				}
@@ -397,12 +415,21 @@ if ( ! class_exists( 'Jetpack_SSO_User_Admin' ) ) :
 						'jetpack-sso-invite-error' => $error, // general error message
 						'_wpnonce'                 => $nonce,
 					);
+
+					$tracking_event_data = array(
+						'success'    => 'false',
+						'error_code' => $error,
+					);
+
+					$body = json_decode( wp_remote_retrieve_body( $response ) );
+					if ( ! empty( $body ) && ! empty( $body->message ) ) {
+						$query_params['jetpack-sso-api-error-message'] = $body->message;
+						$tracking_event_data['error_message']          = $body->message;
+					}
+
 					self::$tracking->record_user_event(
 						$event,
-						array(
-							'success'       => 'false',
-							'error_message' => $error,
-						)
+						$tracking_event_data
 					);
 					return self::create_error_notice_and_redirect( $query_params );
 				}
@@ -937,17 +964,26 @@ if ( ! class_exists( 'Jetpack_SSO_User_Admin' ) ) :
 		 * @return array
 		 */
 		public function jetpack_user_connected_th( $columns ) {
+
+			$tooltip_string = esc_attr__( 'Jetpack SSO allows a seamless and secure experience on WordPress.com. Join millions of WordPress users who trust us to keep their accounts safe.', 'jetpack' );
+
 			wp_enqueue_script(
 				'jetpack-sso-users',
-				plugins_url( 'modules/sso/jetpack-sso-users.js', JETPACK__PLUGIN_FILE ),
+				plugins_url( 'jetpack_vendor/automattic/jetpack-connection/src/sso/jetpack-sso-users.js', JETPACK__PLUGIN_FILE ),
 				array(),
 				JETPACK__VERSION,
 				false
 			);
 
+			wp_add_inline_script(
+				'jetpack-sso-users',
+				"var Jetpack_SSOTooltip = { 'tooltipString': '{$tooltip_string}' }",
+				'before'
+			);
+
 			$columns['user_jetpack'] = sprintf(
-				'<span class="jetpack-sso-invitation-tooltip-icon" role="tooltip" aria-label="%3$s: %1$s" tabindex="0">%2$s [?]<span class="jetpack-sso-invitation-tooltip jetpack-sso-th-tooltip">%1$s</span></span>',
-				esc_attr__( 'Jetpack SSO allows a seamless and secure experience on WordPress.com. Join millions of WordPress users who trust us to keep their accounts safe.', 'jetpack' ),
+				'<span class="jetpack-sso-invitation-tooltip-icon jetpack-sso-status-column" role="tooltip" aria-label="%3$s: %1$s" tabindex="0">%2$s</span>',
+				$tooltip_string,
 				esc_html__( 'SSO Status', 'jetpack' ),
 				esc_attr__( 'Tooltip', 'jetpack' )
 			);