+++ projects/plugins/jetpack/tests/php/sync/test_class.jetpack-sync-themes.php
@@ -5,7 +5,7 @@ use Automattic\Jetpack\Sync\Modules;
 
 require_once ABSPATH . 'wp-admin/includes/class-wp-upgrader.php';
 
-// Mock object requiered for test_theme_update()
+/** Mock object requiered for test_theme_update(). */
 class Dummy_Sync_Test_WP_Upgrader {
 	public $skin;
 
@@ -489,7 +489,8 @@ class WP_Test_Jetpack_Sync_Themes extends WP_Test_Jetpack_Sync_Base {
 	 * Verify that all constants are returned by get_objects_by_id.
 	 */
 	public function test_get_objects_by_id() {
-		$module     = Modules::get_module( 'themes' );
+		$module = Modules::get_module( 'themes' );
+		'@phan-var \Automattic\Jetpack\Sync\Modules\Themes $module';
 		$theme_info = $module->get_objects_by_id( 'theme-info', array() );
 		$this->assertEquals( $module->expand_theme_data(), $theme_info );
 	}