+++ projects/js-packages/components/CHANGELOG.md
@@ -2,6 +2,21 @@
 
 ### This is a list detailing changes for the Jetpack RNA Components package releases.
 
+## [0.53.1] - 2024-05-09
+### Added
+- Added GlobalNotices component and useGlobalNotices hook [#37286]
+
+## [0.53.0] - 2024-05-08
+### Added
+- Social: Added add connection modal [#37211]
+
+### Changed
+- Jetpack Backup: Add a LoadingPlaceholder while waiting for Jetpack Backup price [#37238]
+
+## [0.52.1] - 2024-05-06
+### Changed
+- Updated package dependencies. [#37147] [#37148] [#37160]
+
 ## [0.52.0] - 2024-04-11
 ### Added
 - Added connected card to the A4A plugin. [#36747]
@@ -1008,6 +1023,9 @@
 ### Changed
 - Update node version requirement to 14.16.1
 
+[0.53.1]: https://github.com/Automattic/jetpack-components/compare/0.53.0...0.53.1
+[0.53.0]: https://github.com/Automattic/jetpack-components/compare/0.52.1...0.53.0
+[0.52.1]: https://github.com/Automattic/jetpack-components/compare/0.52.0...0.52.1
 [0.52.0]: https://github.com/Automattic/jetpack-components/compare/0.51.0...0.52.0
 [0.51.0]: https://github.com/Automattic/jetpack-components/compare/0.50.5...0.51.0
 [0.50.5]: https://github.com/Automattic/jetpack-components/compare/0.50.4...0.50.5